tests/test_ai_module.py
import unittest
from ai_module import AIModule

class TestAIModule(unittest.TestCase):
    def setUp(self):
        self.ai_module = AIModule()

    def test_predict_demand(self):
        # Test case 1
        result1 = self.ai_module.predict_demand(historical_data1)
        self.assertEqual(result1, expected_demand1)

        # Test case 2
        result2 = self.ai_module.predict_demand(historical_data2)
        self.assertEqual(result2, expected_demand2)

    def test_detect_anomalies(self):
        # Test case 1
        result1 = self.ai_module.detect_anomalies(data1)
        self.assertEqual(result1, expected_anomalies1)

        # Test case 2
        result2 = self.ai_module.detect_anomalies(data2)
        self.assertEqual(result2, expected_anomalies2)

    # Add more test methods for other AI module functionalities

if __name__ == '__main__':
    unittest.main()